feat(renderOn): Add renderOn configuration option which will cause the view to be rendered when the
renderOn accepts a string, which is the name of the event that will trigger the re-rendering of the

For example, if you want to disable any dates or times that are in the past.
You can $broadcast the event at an interval to disable times in the past (or any other time valid dates change).

closes #166, closes #175