fix(input-directive): input does not remove ng-invalid with valid model value
When the user enters an invalid date (i.e. past date or total non-sense) the input validates and
adds 'ng-invalid' to the input. Then the model is updated to a valid input (i.e. not through the
input but rather the date-picker or an async call returning a value) the directive correctly sets
`ng-valid` on the input. Prior to this fix, the input would not re-validate when changes to the
model happened.

Fix #448